filteration

+++ src/store/products.js
@@ -1,10 +1,8 @@
 import { create } from "zustand";
 
-// Check if window is defined (i.e., we're in a browser environment)
 const isBrowser = typeof window !== "undefined";
 
 const useProductStore = create((set) => {
-  // Check if localStorage is available
   const storedProducts = isBrowser
     ? JSON.parse(localStorage.getItem("products")) || []
     : [];
@@ -14,14 +12,25 @@ const useProductStore = create((set) => {
 
     addToCart: (product) =>
       set((state) => {
-        const updatedProduct = {
-          ...product,
-          quantity: product.quantity ? product?.quantity : 1,
-        };
+        const existingProductIndex = state.products.findIndex(
+          (p) => p.id === product.id
+        );
 
-        const updatedProducts = [...state.products, updatedProduct];
-        localStorage.setItem("products", JSON.stringify(updatedProducts));
-        return { products: updatedProducts };
+        if (existingProductIndex !== -1) {
+          const updatedProducts = [...state.products];
+          updatedProducts[existingProductIndex].quantity +=
+            product.quantity || 1;
+          localStorage.setItem("products", JSON.stringify(updatedProducts));
+          return { products: updatedProducts };
+        } else {
+          const updatedProduct = {
+            ...product,
+            quantity: product.quantity || 1,
+          };
+          const updatedProducts = [...state.products, updatedProduct];
+          localStorage.setItem("products", JSON.stringify(updatedProducts));
+          return { products: updatedProducts };
+        }
       }),
 
     removeFromCart: (productId) =>
